Start demos with pre-loaded settings
We can't change all of Phabricator's settings from the
command-line.  According to https://secure.phabricator.com/T5310, the
best way to pre-load these settings before container startup is to take
a SQL dump of a configured system and load it during container start.

`arc diff` now works out-of-the-box for the `phab` user on demo
container startup. Instructions are in README.md.