add http proxy to the docker network.
This commit introduces a tinyproxy instance to the docker network. This
prevent port collisions, host file modifications, and make interacting
with the docker network from the browser more convenient.

A "firefox-proxy" bash script is also introduced to make running a
firefox instance with a proxied profile easier. This will create a
folder for the profile in the repository root.

BMO currently gives a 500 error when accessed, but that is due to a
bug in the mozilla-conduit/docker-bmo repository. After the bug is
fixed everything should function correctly.