conduit-demo: Update demo BMO environment to match new changes in bmo-extensions git repo
Summary:
Recently, the bmo-extensions development environment underwent some refactoring thanks to Dylan.
It now uses proper separation of all BMO services in their own containers and also uses the upstream
BMO docker images which are maintained.

This bug is to port over the same enhancements to the conduit-demo development environment so that
the two are similar.

In the future, I would like to break this out somehow to allow the two environments to share the
same BMO settings without duplication between the two repos.

Test Plan:
docker-compose up should bring up all necessary BMO related services and accessing http://bmo.test
should look and work as before. The update credentials for BMO are in the README.md.

Reviewers: zalun, davidwalsh

Reviewed By: zalun, davidwalsh

Bug #: 1425087

Differential Revision: https://phabricator.services.mozilla.com/D337