cinnabar: Use modified arcanist in the demo environment
Summary:
To be able to use git-cinnabar in the demo we need to change the
standard library from `musl` to `glibc`. This requires change of the
base image.

Git-cinnabar extension is cloned and installed.

VCT repository is cloned with git-cinnabar into `vct-cinnabar` directory.

We're installing arcanist from the `mozilla-conduit` fork.
Cinnabar flavored Arcanist is aliased and can be used as `cinnabarc`.

Test Plan:
Use the vct-cinnabar and version-control-tools. Run `cinnabarc diff ...` from one and
`arc patch ...` from the other.
To get the full experience you need to make the Phabricator blocking the `arc diff` if VCS of local repository do not match the Phabricator one - apply D1131 on phabricator-extensions

Reviewers: dkl

Reviewed By: dkl

Subscribers: glandium, smacleod

Bug #: 1459922

Differential Revision: https://phabricator.services.mozilla.com/D1211