suite: Add Transplant service
Summary:
To make the Suite complete we need to be able to land the diff within
the local development environment. This patch is adding the Transplant
application cluster.

The only functionalities which do require the internet connection is the
Lando user authentication via Auth0 and storing diffs in an AWS S3 bucket.

- Added a script to (re)clone the test repositories in `local-dev` service.
- Refactored `docker-compose.yml` and organized the services into logical groups.
- Added `docker-compose.transplant.yml` for local Transplant development.
- Renamed `VCT` repository to `TESTREPO` in Phabricator.
- Added a "landing step by step" section to the readme.

Test Plan:
Run the Successful landing step by step.
If you'd like to read the Readme in a rendered state please load the https://github.com/zalun/suite/blob/transplant/README.md

Reviewers: dkl, glob

Reviewed By: dkl, glob

Bug #: 1465003

Differential Revision: https://phabricator.services.mozilla.com/D1826