No bug - move virtual environment instructions above `moz-phab` r?dkl
When running through these docs, I was briefly hung up on the `moz-phab` step since it wasn't on my `PATH`.
I had to come back to the README to see that a virtual environment was necessary, but it wasn't mentioned until
later in the doc. By moving the steps around, it makes following the instructions sequentially a little easier.

Differential Revision: https://phabricator.services.mozilla.com/D56774