Bug 1606383 - update docker-compose version to 3.4
The only breaking change is that "volumes_from" no longer exists. So, I had to enumerate the volumes of each
image + service config and encode them where a ".test" service would have done "volumes_from".

In the "docker-compose.*.yml" files where the shared volume was a file/dir on the host file system, I separately defined
the volume in the top-level "volumes" section and had the associated services refer to that.
I used http://blog.code4hire.com/2018/06/define-named-volume-with-host-mount-in-the-docker-compose-file/ as a reference here.

Differential Revision: https://phabricator.services.mozilla.com/D58418