docker-compose: don't export internal ports (bug 1935046)
The `ports` directive exposes the specified service's port on the local
host. Most of the ports in the suite stack only need to be reachable
within the stack, with only the tinyproxy port exposed more widely.

Removing this prevents port collision when running multiple stacks using
common ports (e.g., 80, 5432, ...).