suite: wrapper for docker compose, with overrides support (#75)
We already have docker-compose.override.yml in the .gitignore. It's a
good place to add temporary local overrides that shouldn't be committed.

Using the override, however, makes the `docker compose` a bit clumsy.
The `./suite` script can be used as a drop-in replacement of the system
`docker compose`, and adds the necessary command line arguments so
the overrides are honoured.